# regalloc: removes map use for less memory pressure