This yields better perf on compilation as you see in the following bench result for 
compiling the wazero CLI compiled as a Wasm binary:
```
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero
               │  old.txt   │           new.txt           │
               │   sec/op   │   sec/op    vs base         │
Compilation-10   3.467 ± 1%   3.523 ± 4%  ~ (p=0.073 n=7)

               │   old.txt    │              new.txt               │
               │     B/op     │     B/op      vs base              │
Compilation-10   396.2Mi ± 0%   392.7Mi ± 0%  -0.89% (p=0.001 n=7)

               │   old.txt   │              new.txt              │
               │  allocs/op  │  allocs/op   vs base              │
Compilation-10   718.9k ± 0%   693.2k ± 0%  -3.58% (p=0.001 n=7)
```